+++ HeadFirstDesignPatterns/src/util/JavaLog.java
@@ -1,4 +1,4 @@
-package tuil;
+package util;
 
 
 public class JavaLog {
